requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
python-multipart==0.0.6
docker==6.1.3
redis==5.0.1
celery==5.3.4
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1